third_party/blink/web_tests/fast/replaced/iframe-with-percentage-height-within-table-with-table-cell-ignore-height-expected.txt

This test checks that iframe with percentage height within table cell ignores the table cell height attribute (strict mode).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getHeight('iframe-100') is window.innerHeight
PASS successfullyParsed is true

TEST COMPLETE
